🚜 폴더명 변경
