+++ src/components/Terms/Markdown/MarkdownEditor.tsx
@@ -1,5 +1,5 @@
 import dynamic from 'next/dynamic'
-import React, { useCallback, useRef, FC } from 'react'
+import React, { useState, useCallback, useRef, FC, useEffect } from 'react'
 import { Editor as EditorType, EditorProps } from '@toast-ui/react-editor'
 import { TuiEditorWithForwardedProps } from './TuiEditorWrapper'
 
@@ -11,7 +11,6 @@ const Editor = dynamic<TuiEditorWithForwardedProps>(
   () => import('./TuiEditorWrapper'),
   { ssr: false },
 )
-
 const EditorWithForwardedRef = React.forwardRef<
   EditorType | undefined,
   EditorPropsWithHandlers
@@ -22,8 +21,8 @@ const EditorWithForwardedRef = React.forwardRef<
 interface Props extends EditorProps {
   onChange?(value: string | boolean): void
   valueType?: 'markdown' | 'html'
+  setIsContentEmpty?: (isContentEmpty: boolean) => void
   setMarkdownContent?: (content: any) => void
-  isMarkdownContentEmpty?: boolean
 }
 
 const Markdown: FC<Props> = (props, value) => {
@@ -38,7 +37,13 @@ const Markdown: FC<Props> = (props, value) => {
   const editorRef = useRef<EditorType>()
   const handleChange = useCallback(() => {
     const instance = editorRef.current.getInstance()
-    const valueType = props.valueType ?? 'markdown'
+    if (instance.getMarkdown().length < 1) {
+      props.setIsContentEmpty(true)
+      return
+    } else {
+      props.setIsContentEmpty(false)
+    }
+    const valueType = props.valueType || 'markdown'
     props.setMarkdownContent(
       valueType === 'markdown' ? instance.getHTML() : instance.getMarkdown(),
     )
@@ -50,9 +55,9 @@ const Markdown: FC<Props> = (props, value) => {
         {...props}
         initialValue={initialValue}
         previewStyle={previewStyle ?? 'vertical'}
-        height={height ?? '550px'}
-        initialEditType={initialEditType ?? 'markdown'}
-        useCommandShortcut={useCommandShortcut ?? true}
+        height={height || '550px'}
+        initialEditType={initialEditType || 'markdown'}
+        useCommandShortcut={useCommandShortcut || true}
         ref={editorRef}
         onChange={handleChange}
       />
