+++ src/utils/fetcher.ts
@@ -1,11 +1,10 @@
 import axios from 'axios'
 
-const fetcher = (url: string) => {
-  return axios
-    .get(`${process.env.NEXT_PUBLIC_API_URL}/${url}`, {
+const fetcher = (url: string) =>
+  axios
+    .get(url, {
       withCredentials: true,
     })
     .then((response) => response.data)
-}
 
 export default fetcher
