+++ src/utils/getDataFromLocalStorage.ts
@@ -1,6 +1,4 @@
 const getDataFromLocalStorage = (key: string) => {
-  if (typeof window !== 'undefined') {
-    return localStorage.getItem(key)
-  }
+  return localStorage.getItem(key)
 }
 export default getDataFromLocalStorage
